Iterate on SSL_CREDENTIAL_set_must_match_issuer a bit
First, simplify the API a bit:
- Just take a boolean param rather than having both set and clear
  functions.
- Unless we need it, no need to bother with a getter. We generally
  assume that the caller knows what they configured.

Next, expand on the docs and move it with other credential APIs, not
SSL_PRIVATE_KEY_METHOD.

Finally, fix a bug and test this in runner: the TLS 1.2 handshake forgot
to check the issuer, which meant that it assumed all credentials were
viable. Fix this and add tests to cover it all. In doing so, this pulls
in the MustMatchIssuer runner plumbing out of
https://boringssl-review.googlesource.com/c/boringssl/+/73087 to land a
little sooner.

Also test that issuer matching works with delegated credentials. May as
well.

Change-Id: I22aee148dd81fb9804d80b4243b68a5ecdead480
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/76708
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: David Benjamin <davidben@google.com>