aes-gcm-avx10-x86_64.pl: fold _ghash_mul_step into _ghash_mul
Fold _ghash_mul_step into _ghash_mul, since the support for interleaving
a single-vector GHASH multiplication with other instructions is not
used.  It is used in the Linux port (for _aes_gcm_final), but it is not
used in the BoringSSL port.

No change to the generated assembly code.

Change-Id: I5a65ee490e814ca6390c4b968ab135091c344a98
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77167
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: David Benjamin <davidben@google.com>