Move some exponentation functions out of the FIPS module
BN_exp is never used in cryptography.

We never call BN_mod_exp from the library and instead call the specific
implementations directly.

We never call BN_mod_exp_mont_word. Looks like this was originally added
for optimized small generator operations in Diffie-Hellman, but we
dropped the optimizations. We consider (non-EC) Diffie-Hellman to be a
legacy algorithm.

BN_mod_exp2_mont is used in DSA, but we dropped the optimizations
(non-EC) DSA is also a legacy algorithm and DSA is not in the FIPS
module.

Change-Id: I1cf2975a834f1da12f4e7623339739de151a3e1e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77447
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: Bob Beck <bbe@google.com>
Auto-Submit: David Benjamin <davidben@google.com>