Add support for C2PA validation modes
This permits chains to be validated when the intended
use is to validate a C2PA time stamp signer or C2PA
manifest signer.

As in the RCS MLS case, further C2PA specific checks
for the purpose intended may be needed on the certificate
after verify is run.

Bug: 400963456
Change-Id: I22bfe6b2d3ad3cfd939a2c5a13debb5f0d3ccbb3
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77487
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: David Benjamin <davidben@google.com>