Remove bn_div_words from bn-586.pl
This function is unused. While there is a bn_div_words symbol, it's
static and we never actually use the assembly-defined one.

Not using it is potentially a missed optimization, but only for 32-bit
x86 on MSVC because for GCC/Clang we have inline assembly. (It is hard
for the compiler to emit the instruction itself because it cannot prove
that the division fits.)

Change-Id: Ia7ecd28f26143521343dd516a07eb3adf9c4378e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77627
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>