Mark some transitive includes as IWYU export.
This reflects some assumptions we have on our headers:

- <openssl/foo.h> pulls in <openssl/base.h>. In particular, the
  canonical FOO typedefs for foo_st all live forward declared in
  <opessl/base.h>, but including <openssl/foo.h> is sufficient to use
  FOO.

- <openssl/base.h> pulls in <stdint.h> and <stddef.h> so we don't have
  to keep including it.

Add IWYU exports to reflect this so that clang's include cleaner gets
less upset. It's a bit of a blunt instrument because it also means that
<openssl/foo.h> lets you use the forward-declared BAR typedef, and
downstream projects might still prefer to explicit include <stdint.h>
when they use it (we use it so much that it would be too much), but I
think this is fine.

NB: By adding these, we're essentially promising we won't include those
transitive includes because downstream code might accidentally start
relying on it.

Change-Id: I705fe6d1026fbd302ed0f070a0cf7658a70af8ef
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77687
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>