Fix the link order of various dependencies
CMake expects you to provide your link lines in the right order for
platforms without rescanning linkers. When they're not in the right,
it'll preserve your order but duplicate transitive dependencies. That
is, if:

    A -> C B
    B -> C

The final link line will be:

    A C B C

Wheras if A wrote B C then there would be no duplication. Newer macOS
toolchains (which do not need the duplication) seem to warn on duplicate
libraries, which is how I noticed this.

That said, this is not actually sufficient to avoid duplication and thus
the warning. Consider:

   A -> B D
   B -> C D

CMake always lists direct dependencies before transitive ones, so the
result will be:

   A B D C D

decrepit_test triggers this because decrepit_test does not directly
depend on ssl but decrepit does. It's a bit awkward to have to list it
again, but adding it avoids this issue.

Newer CMakes (3.31) now know that:
1. macOS rescans dependencies so the dependencies don't have to be in
   order.
2. macOS has this warning so it should dedup things.

However, even updating to 3.31 isn't sufficient because CMake keys all
behavior changes on cmake_minimum_required. I didn't set the target
policy version because I figure testing it at 3.16's behavior is
probably useful for the sake of keeping 3.16 hopefully working.

Change-Id: Ibb006eefbbb23f1899ef91277465d31fa8202d2e
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/77747
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>