Revert "Remove Karatsuba multiplication in BIGNUM"
This reverts commit 89eb6ddbf52887631ca45ad2242862515cd698ab. Sadly,
exactfloat, a decidedly non-cryptographic use case, uses BN_mul with
large enough inputs that Karatsuba is actually a load-bearing
optimization. (They're also why we need to support allocating giant
BIGNUMs.)

This CL just reverts the change for now. We should revise thresholds and
rearrange code so that this code is not reachable from any of the
cryptographic code. From there, we can revert the work to make it
constant-time, which will be better from exactfloat and also remove some
complexity.

Bug: 406497222
Change-Id: I08b6d12e19c2a6ae741ac490d81cc534ba260145
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/78047
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: Bob Beck <bbe@google.com>