Fix comment on bn_div_rem_words and use MSVC instrinsic
We write in the comment that it's a compiler bug that GCC and Clang
won't emit the DIV instruction. That dates to
https://boringssl-review.googlesource.com/7127, which was an elaborated
version of an existing OpenSSL comment:

      - GNU C generates a call to a function (__udivdi3 to be exact)
        in reply to ((((BN_ULLONG)n0)<<BN_BITS2)|n1)/d0 (I fail to
        understand why...);

Turns out the compiler is right. If the quotient doesn't fit in a single
word, the DIV instruction traps. A truncated BN_ULLONG / BN_ULONG is
obligated to output the truncated quotient. (Ideally the compiler would
figure it out with bounds on the numerator, but it's hard to figure that
out portably and it doesn't seem to figure it out.) Rewrite the comment
accordingly.

Also, while MSVC doesn't support GCC inline assembly, it has intrinsics
for this operation. We may as well use them, since the operation we'd
otherwise try to do (BN_ULLONG / BN_ULONG) isn't great anyway. This
gives a small speed boost on MSVC. (We use BN_div in computing public
BN_MONT_CTX, which is on path to RSA verification.) Not huge, but costs
very little.

Before:
Did 125750 RSA 2048 verify (fresh key) operations in 3015000us (41708.1 ops/sec)
After:
Did 130250 RSA 2048 verify (fresh key) operations in 3031000us (42972.6 ops/sec)

Change-Id: I3e0a9a1ab31d2b9de7bb757837b2d33fc7f81941
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/78147
Reviewed-by: Bob Beck <bbe@google.com>
Commit-Queue: David Benjamin <davidben@google.com>
Auto-Submit: David Benjamin <davidben@google.com>