Replace sscanf() with strtoull() in OPENSSL_ia32cap parser
This simplifies the code a bit while still accepting the same syntax as
previously: decimal or hexadecimal (hex must be prefixed with exactly
"0x", case sensitive), potentially clearing the corresponding bits if
prefixed with "~".

An explicit base (either 10 or 16) is passed to strtoull() to ensure
that zero-prefixed decimal numbers are not parsed as octal.

As a side effect, this also avoids the stdio dependency, which is
helpful for bare-metal environments.

Change-Id: I2b250161eb154c5e0bbc7629b695f573f4b21733
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/78367
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: Adam Langley <agl@google.com>