Test non-NEON codepaths with a OPENSSL_NO_STATIC_NEON_FOR_TESTING
This removes the blocker on updating the NDK. The actual test-only build
flag isn't amazing, but it's much simpler than assembling a custom Arm
toolchain. Time spent on that is probably better spent on removing the
assembly in question, as it'll only be less and less relevant.

Note that, although this CL changes code and main.star at the same
time, LUCI is not capable of updating them atomically, so we'll still
run the old main.star against the new code.

Fixed: 42290318
Change-Id: Ia08350284d47cd5767aa6fd15a2e44443b90ae18
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/79148
Reviewed-by: Adam Langley <agl@google.com>
Commit-Queue: David Benjamin <davidben@google.com>