Remove the SHA-1 default for PSS with EVP_PKEY
Nothing seems to rely on this for PSS, which is good as SHA-1 is not
safe for use with PSS.

The OAEP default is sadly harder to remove. Node depends on it. However
SHA-1 in OAEP is less pressing, so this is probably okay to leave alone
for now, though not ideal.

Update-Note: PSS with the EVP APIs will no longer work without an
explicit choice of hash function. Previously the default was SHA-1, a
broken hash algorithm. Callers can restore the use of SHA-1 with
EVP_PKEY_CTX_set_signature_md.

Bug: 389733571
Change-Id: I8844764989d170d4b0b435367f0b15d5b892d96d
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/80008
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Adam Langley <agl@google.com>
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: Adam Langley <agl@google.com>